Fix missing nose
Certain combinations of FG clients on a Multiplayer network will fail to
transmit all the properties correctly due to a packet size limit in the
1.0 version of the MP protocol. As a result, the "blunt-nose" property
is sometimes missing, and this leads to the DHC6 being rendered without
any nose at all. This patch fixes the nose selection animation such that
in the absence of a blunt-nose property, it will default to the blunt
nose, rather than no nose at all.